Time: 4 ms (78.89%), Space: 25.8 MB (99.66%) - LeetHub

+++ 3158-find-the-xor-of-numbers-which-appear-twice/3158-find-the-xor-of-numbers-which-appear-twice.cpp
@@ -4,7 +4,7 @@ public:
         bitset<51>vis;
         for(auto&it:nums)
             vis[it].flip();
-        int xr=0;
+        short xr=0;
         for(auto&it:nums)
             if(!vis[it])
                 xr^=it,vis[it].flip();
